Move onMouseLeave event handler to ul
Active styling is already being reset from previously active li elements in the onMouseOver event handler (when moving
to a neighbouring li element). However, the sole purpose for adding the onMouseLeave event handler was to remove active
styling when the user navigates away from all the li elements.
Placing this on the li resulted in it being called each time the mouse left an li element, when really it should
only be called when leaving all the li elements (the ul element).